+++ tests/core/tracer/test_in_memory_tracer.py
@@ -99,7 +99,10 @@ def test_task_span_records_error_value() -> None:
         raise ValueError("my bad, sorry")
 
     assert isinstance(tracer.entries[0], InMemoryTaskSpan)
-    error = tracer.entries[0].output
+    error_log = tracer.entries[0].entries[0]
+    assert isinstance(error_log, LogEntry)
+
+    error = error_log.value
     assert isinstance(error, ErrorValue)
     assert error.message == "my bad, sorry"
     assert error.error_type == "ValueError"
@@ -182,7 +185,7 @@ def set_env(name: str, value: str | None) -> Iterator[None]:
         os.environ.update(old_environ)
 
 
-def test_in_memory_tracer_trace_viewer_doesnt_crash_if_it_cant_reach_document_index() -> (
+def test_in_memory_tracer_submit_to_trace_viewer_doesnt_crash_if_it_cant_reach_the_trace_viewer() -> (
     None
 ):
     # note that this test sets the environment variable, which might